+++ pom.xml
@@ -1,8 +1,8 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
   <modelVersion>4.0.0</modelVersion>
-  <groupId>com.pwdd.httpServer</groupId>
-  <artifactId>httpServer</artifactId>
+  <groupId>com.pwdd.server</groupId>
+  <artifactId>server</artifactId>
   <packaging>jar</packaging>
   <version>0.1</version>
   <name>httpServer</name>
@@ -17,7 +17,7 @@
           <archive>
             <manifest>
               <addClasspath>true</addClasspath>
-              <mainClass>com.pwdd.httpServer.HTTPServer</mainClass>
+              <mainClass>com.pwdd.server.HTTPServer</mainClass>
             </manifest>
           </archive>
         </configuration>
