--- src/main/java/com/pwdd/httpServer/HTTPServer.java
@@ -1,9 +1,0 @@
-package com.pwdd.httpServer;
-
-import com.pwdd.httpServer.connection.ServerRunner;
-
-class HTTPServer {
-  public static void main( String[] args ) {
-    ServerRunner.start(args);
-  }
-}
