--- src/main/java/com/pwdd/httpServer/responders/IResponder.java
@@ -1,9 +1,0 @@
-package com.pwdd.httpServer.responders;
-
-public interface IResponder {
-  boolean canRespond(String uri);
-
-  byte[] header(String date);
-
-  byte[] body(String uri);
-}
