+++ src/main/java/com/pwdd/server/HTTPServer.java
@@ -1,0 +1,9 @@
+package com.pwdd.server;
+
+import com.pwdd.server.server.ServerRunner;
+
+public class HTTPServer {
+  public static void main( String[] args ) {
+    ServerRunner.start(args);
+  }
+}
