+++ src/main/java/com/pwdd/server/responders/IResponder.java
@@ -1,0 +1,11 @@
+package com.pwdd.server.responders;
+
+public interface IResponder {
+  String CRLF = "\r\n";
+
+  boolean canRespond(String fullURI);
+
+  byte[] header(String date);
+
+  byte[] body(String fullURI);
+}
