+++ src/test/java/com/pwdd/server/helpers/Helpers.java
@@ -1,0 +1,11 @@
+package com.pwdd.server.helpers;
+
+import java.nio.charset.StandardCharsets;
+
+public final class Helpers {
+  private Helpers() {}
+
+  public static String bytesToString(byte[] in) {
+    return new String(in, StandardCharsets.UTF_8);
+  }
+}
