+++ src/test/java/com/pwdd/server/mocks/filesystem/monkey.gif
